+++ src/ccip/BurnMintERC677Helper.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: BUSL-1.1
 pragma solidity ^0.8.19;
 
-import {BurnMintERC677} from "@chainlink/contracts-ccip/src/v0.8/shared/token/ERC677/BurnMintERC677.sol";
+import {BurnMintERC677} from "@chainlink/contracts/src/v0.8/shared/token/ERC677/BurnMintERC677.sol";
 
 /// @title BurnMintERC677Helper
 /// @notice This contract extends the functionality of the BurnMintERC677 token contract to include a `drip` function that mints one full token to a specified address.
@@ -13,10 +13,7 @@ contract BurnMintERC677Helper is BurnMintERC677 {
      * @param name - The name of the token.
      * @param symbol - The symbol of the token.
      */
-    constructor(
-        string memory name,
-        string memory symbol
-    ) BurnMintERC677(name, symbol, 18, 0) {}
+    constructor(string memory name, string memory symbol) BurnMintERC677(name, symbol, 18, 0) {}
 
     /**
      * @notice Mints one full token (1e18) to the specified address.